requirements.txt
# Web interface
flask>=3.0.0
gunicorn>=22.0.0          # Production WSGI server (Azure App Service)

# Web search (DuckDuckGo – no API key required)
ddgs>=1.0.0

# Azure AI – Agents SDK & project client
azure-ai-projects>=1.0.0
azure-ai-agents>=1.0.0

# Azure identity (DefaultAzureCredential – works with az login, Managed Identity, etc.)
azure-identity>=1.19.0

# HTTP utilities (for any supplemental URL fetching)
httpx>=0.27.0

# Output rendering
jinja2>=3.1.0          # HTML scorecard template rendering
markdown>=3.6          # Optional: convert Markdown to HTML standalone

# CLI / UX
rich>=13.7.0           # Pretty console output
click>=8.1.0           # CLI argument parsing

# Environment loading
python-dotenv>=1.0.0
